+++ server/src/main/java/org/elasticsearch/timeseries/support/TimeSeriesMetrics.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.search.tsdb;
+package org.elasticsearch.timeseries.support;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
